Refactor ServiceApp by moving app into a single file
